requirements.txt
Flask==3.1.2
flask-cors==6.0.1
praw==7.8.1
pandas==2.3.3
numpy==2.2.6
openai==2.3.0
pinecone-client==6.0.0
langchain==0.3.27
langchain-openai==0.3.35
langchain-pinecone==0.2.12
langchain-community==0.3.31
nltk==3.9.2
vaderSentiment==3.3.2
python-dotenv==1.1.1
requests==2.32.5
tqdm==4.67.1
PyYAML==6.0.3
